+++ hutool-core/src/main/java/cn/hutool/core/bean/copier/BeanCopier.java
@@ -244,25 +244,27 @@ public class BeanCopier<T> implements Copier<T>, Serializable {
 				continue;
 			}
 
-			Type valueType = (null == setterMethod) ? TypeUtil.getType(field) : TypeUtil.getFirstParamType(setterMethod);
-			if (valueType instanceof ParameterizedType) {
-				// 参数为泛型参数类型，解析对应泛型类型为真实类型
-				ParameterizedType tmp = (ParameterizedType) valueType;
-				Type[] actualTypeArguments = tmp.getActualTypeArguments();
+			// 获取目标字段真实类型
+			Type fieldType = (null == setterMethod) ? TypeUtil.getType(field) : TypeUtil.getFirstParamType(setterMethod);
+			if (fieldType instanceof ParameterizedType) {
+				// 字段类型为泛型参数类型，解析对应泛型类型为真实类型，类似于List<T> a
+				final ParameterizedType fieldParameterizedType = (ParameterizedType) fieldType;
+				Type[] actualTypeArguments = fieldParameterizedType.getActualTypeArguments();
 				if (TypeUtil.hasTypeVeriable(actualTypeArguments)) {
 					// 泛型对象中含有未被转换的泛型变量
-					actualTypeArguments = TypeUtil.getActualTypes(this.destType, field.getDeclaringClass(), tmp.getActualTypeArguments());
+					actualTypeArguments = TypeUtil.getActualTypes(this.destType, field.getDeclaringClass(), fieldParameterizedType.getActualTypeArguments());
 					if (ArrayUtil.isNotEmpty(actualTypeArguments)) {
 						// 替换泛型变量为实际类型
-						valueType = new ParameterizedTypeImpl(actualTypeArguments, tmp.getOwnerType(), tmp.getRawType());
+						fieldType = new ParameterizedTypeImpl(actualTypeArguments, fieldParameterizedType.getOwnerType(), fieldParameterizedType.getRawType());
 					}
 				}
-			} else if (valueType instanceof TypeVariable) {
-				// 参数为泛型，查找其真实类型（适用于泛型方法定义于泛型父类）
-				valueType = TypeUtil.getActualType(this.destType, field.getDeclaringClass(), valueType);
+			} else if (fieldType instanceof TypeVariable) {
+				// 字段类型为泛型，查找其真实类型（适用于泛型方法定义于泛型父类），类似于T a
+				fieldType = TypeUtil.getActualType(this.destType, field.getDeclaringClass(), fieldType);
 			}
 
-			value = valueProvider.value(providerKey, valueType);
+			//
+			value = valueProvider.value(providerKey, fieldType);
 			if (null == value && copyOptions.ignoreNullValue) {
 				continue;// 当允许跳过空时，跳过
 			}