+++ hutool-core/src/test/java/cn/hutool/core/util/TypeUtilTest.java
@@ -57,4 +57,9 @@ public class TypeUtilTest {
 		}
 	}
 
+	@Test
+	public void getActualTypesTest(){
+
+	}
+
 }